feat: improve review performance and data integrity
- Add safe long-to-int conversion for review counts
- Add organization_id indexes for faster multi-tenant queries
- Create ReviewDto for better API structure
- Fix potential overflow in review count display

+++ src/main/java/com/reputul/backend/controllers/DashboardController.java
@@ -54,7 +54,7 @@ public class DashboardController {
                 .address(b.getAddress())
                 .reputationScore(b.getReputationScore())
                 .badge(b.getBadge())
-                .reviewCount((int) reviewRepo.countByBusinessId(b.getId()))
+                .reviewCount(Math.toIntExact(reviewRepo.countByBusinessId(b.getId())))
                 .reviewPlatformsConfigured(b.getReviewPlatformsConfigured())
                 .build()
         ).toList();
